+++ test/renderers-reference.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const linz = require('linz');
+const { collections } = require('./lib');
 
 beforeAll((done) => {
 
@@ -50,10 +51,19 @@ beforeAll((done) => {
 
 }, 10000);
 
-afterEach(async () => {
+beforeEach(async (done) => {
 
-    await linz.api.model.get('org').collection.drop();
-    await linz.api.model.get('user').collection.drop();
+    let existingCollections = await collections();
+
+    if (existingCollections.includes['org']) {
+        await linz.api.model.get('org').collection.drop();
+    }
+
+    if (existingCollections.includes['user']) {
+        await linz.api.model.get('user').collection.drop();
+    }
+
+    return done();
 
 });
